기존 로그인 컴포넌트를 리팩토링하여 Firebase 인증 로직을 훅으로 분리하고, 사용자 상태 및 인증 UI를 별도의 컴포넌트로 분리했습니다. 또한, 디렉토리 및 파일 링크 처리 로직을 개선하고, HTML 콘텐츠를 React 컴포넌트로 변환하는 기능을 추가했습니다.
