+++ app/types/pagePermissions.ts
@@ -33,12 +33,12 @@ export const pagePermissions: PagePermission[] = [
   },
   // 어드민 권한만 접근 가능
   {
-    path: '/1. 일지/*',
+    path: '/1. 일지*',
     allowedRoles: [UserRole.ADMIN],
     isPublic: false,
   },
   {
-    path: '/3. 회사/*',
+    path: '/3. 회사*',
     allowedRoles: [UserRole.ADMIN],
     isPublic: false,
   },
@@ -58,24 +58,24 @@ export const pagePermissions: PagePermission[] = [
     isPublic: false,
   },
   {
-    path: '/97. 보안 폴더/*',
+    path: '/97. 보안 폴더*',
     allowedRoles: [UserRole.ADMIN],
     isPublic: false,
   },
   {
-    path: '/99. 일기/*',
+    path: '/99. 일기*',
     allowedRoles: [UserRole.ADMIN],
     isPublic: false,
   },
 
   // 어드민, 인증된 사용자만 접근 가능
   {
-    path: '/8. 루틴/*',
+    path: '/8. 루틴*',
     allowedRoles: [UserRole.ADMIN, UserRole.VERIFIED],
     isPublic: false,
   },
   {
-    path: '/98. 미분류/*',
+    path: '/98. 미분류*',
     allowedRoles: [UserRole.ADMIN],
     isPublic: false,
   },
